Proper Block Checks for Company and Tests
