+++ crates/bcr-ebill-core/src/protocol/blockchain/company/mod.rs
@@ -1,28 +1,19 @@
+use crate::protocol::crypto::BcrKeys;
+use crate::protocol::{BlockId, EmailIdentityProofData, SignedIdentityProof};
+
 use super::Result;
-use super::bill::BillOpCode;
 use super::{Block, Blockchain};
-use crate::protocol::BlockId;
-use crate::protocol::City;
-use crate::protocol::Country;
-use crate::protocol::Date;
-use crate::protocol::Email;
-use crate::protocol::Identification;
-use crate::protocol::Name;
-use crate::protocol::SchnorrSignature;
-use crate::protocol::Sha256Hash;
-use crate::protocol::Timestamp;
-use crate::protocol::base::identity_proof::{EmailIdentityProofData, SignedIdentityProof};
-use crate::protocol::blockchain::{Error, borsh_to_json_value};
-use crate::protocol::crypto::{self, BcrKeys};
-use crate::protocol::{File, OptionalPostalAddress, PostalAddress};
-use bcr_common::core::BillId;
 use bcr_common::core::NodeId;
-use bitcoin::base58;
-use borsh::{from_slice, to_vec};
 use borsh_derive::{BorshDeserialize, BorshSerialize};
-use log::error;
-use secp256k1::{PublicKey, SecretKey};
 use serde::{Deserialize, Serialize};
+
+pub mod block;
+pub mod chain;
+pub mod validation;
+
+pub use block::CompanyBlock;
+pub use block::CompanyBlockPayload;
+pub use chain::CompanyBlockchain;
 
 #[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub enum CompanyOpCode {
@@ -36,897 +27,14 @@ pub enum CompanyOpCode {
     IdentityProof,
 }
 
-#[derive(BorshSerialize)]
-pub struct CompanyBlockDataToHash {
-    company_id: NodeId,
-    id: BlockId,
-    plaintext_hash: Sha256Hash,
-    previous_hash: Sha256Hash,
-    data: Vec<u8>, // encrypted block data
-    timestamp: Timestamp,
-    #[borsh(
-        serialize_with = "crate::protocol::serialization::serialize_pubkey",
-        deserialize_with = "crate::protocol::serialization::deserialize_pubkey"
-    )]
-    public_key: PublicKey,
-    signatory_node_id: NodeId,
-    op_code: CompanyOpCode,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub enum SignatoryType {
-    Solo,
-}
-
-/// Structure for the block data of a company block
-///
-/// - `data` contains the actual data of the block, encrypted using the company's pub key
-/// - `key` is optional and if set, contains the company private keys encrypted by an identity
-///   pub key (e.g. for CreateCompany the creator's and InviteSignatory the signatory's)
-#[derive(BorshSerialize, BorshDeserialize, Debug, Clone, PartialEq)]
-pub struct CompanyBlockData {
-    data: Vec<u8>,
-    // The encrypted, base58 encoded SecretKey
-    key: Option<String>,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanyBlock {
+#[derive(Debug, Clone)]
+pub struct CompanyValidateActionData {
+    pub blockchain: CompanyBlockchain,
     pub company_id: NodeId,
-    pub id: BlockId,
-    pub plaintext_hash: Sha256Hash,
-    pub hash: Sha256Hash,
-    pub timestamp: Timestamp,
-    pub data: Vec<u8>, // encrypted block data
-    #[borsh(
-        serialize_with = "crate::protocol::serialization::serialize_pubkey",
-        deserialize_with = "crate::protocol::serialization::deserialize_pubkey"
-    )]
-    pub public_key: PublicKey,
-    pub signatory_node_id: NodeId,
-    pub previous_hash: Sha256Hash,
-    pub signature: SchnorrSignature,
-    pub op_code: CompanyOpCode,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone)]
-pub struct CompanyBlockPlaintextWrapper {
-    pub block: CompanyBlock,
-    pub plaintext_data_bytes: Vec<u8>,
-}
-
-impl CompanyBlockPlaintextWrapper {
-    /// This is only used for dev mode
-    pub fn to_json_text(&self) -> Result<String> {
-        let mut serialized =
-            serde_json::to_value(&self.block).map_err(|e| Error::JSON(e.to_string()))?;
-
-        let block_data: serde_json::Value = match self.block.op_code() {
-            CompanyOpCode::Create => {
-                borsh_to_json_value::<CompanyCreateBlockData>(&self.plaintext_data_bytes)?
-            }
-            CompanyOpCode::Update => {
-                borsh_to_json_value::<CompanyUpdateBlockData>(&self.plaintext_data_bytes)?
-            }
-            CompanyOpCode::InviteSignatory => {
-                borsh_to_json_value::<CompanyInviteSignatoryBlockData>(&self.plaintext_data_bytes)?
-            }
-            CompanyOpCode::SignatoryAcceptInvite => borsh_to_json_value::<
-                CompanySignatoryAcceptInviteBlockData,
-            >(&self.plaintext_data_bytes)?,
-            CompanyOpCode::SignatoryRejectInvite => borsh_to_json_value::<
-                CompanySignatoryRejectInviteBlockData,
-            >(&self.plaintext_data_bytes)?,
-            CompanyOpCode::RemoveSignatory => {
-                borsh_to_json_value::<CompanyRemoveSignatoryBlockData>(&self.plaintext_data_bytes)?
-            }
-            CompanyOpCode::SignCompanyBill => {
-                borsh_to_json_value::<CompanySignCompanyBillBlockData>(&self.plaintext_data_bytes)?
-            }
-            CompanyOpCode::IdentityProof => {
-                borsh_to_json_value::<CompanyIdentityProofBlockData>(&self.plaintext_data_bytes)?
-            }
-        };
-
-        if let Some(obj) = serialized.as_object_mut() {
-            obj.insert("data".to_string(), block_data);
-        } else {
-            return Err(Error::JSON(
-                "Block didn't serialize to JSON object".to_string(),
-            ));
-        }
-        serde_json::to_string(&serialized).map_err(|e| Error::JSON(e.to_string()))
-    }
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanyCreateBlockData {
-    pub id: NodeId,
-    pub name: Name,
-    pub country_of_registration: Option<Country>,
-    pub city_of_registration: Option<City>,
-    pub postal_address: PostalAddress,
-    pub email: Email,
-    pub registration_number: Option<Identification>,
-    pub registration_date: Option<Date>,
-    pub proof_of_registration_file: Option<File>,
-    pub logo_file: Option<File>,
-    pub creation_time: Timestamp,
-    pub creator: NodeId,
-}
-
-#[derive(
-    BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, Default, PartialEq,
-)]
-pub struct CompanyUpdateBlockData {
-    pub name: Option<Name>,
-    pub email: Option<Email>,
-    pub postal_address: OptionalPostalAddress,
-    pub country_of_registration: Option<Country>,
-    pub city_of_registration: Option<City>,
-    pub registration_number: Option<Identification>,
-    pub registration_date: Option<Date>,
-    pub logo_file: Option<File>,
-    pub proof_of_registration_file: Option<File>,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanySignCompanyBillBlockData {
-    pub bill_id: BillId,
-    pub block_id: BlockId,
-    pub block_hash: Sha256Hash,
-    pub operation: BillOpCode,
-    #[borsh(
-        serialize_with = "crate::protocol::serialization::serialize_optional_privkey",
-        deserialize_with = "crate::protocol::serialization::deserialize_optional_privkey"
-    )]
-    pub bill_key: Option<SecretKey>,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanyInviteSignatoryBlockData {
-    pub invitee: NodeId,
-    pub inviter: NodeId,
-    pub t: SignatoryType,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanySignatoryAcceptInviteBlockData {
-    pub accepter: NodeId,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanySignatoryRejectInviteBlockData {
-    pub rejecter: NodeId,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanyRemoveSignatoryBlockData {
-    pub removee: NodeId,
-    pub remover: NodeId,
-}
-
-#[derive(BorshSerialize, BorshDeserialize, Serialize, Deserialize, Debug, Clone, PartialEq)]
-pub struct CompanyIdentityProofBlockData {
-    pub proof: SignedIdentityProof,
-    pub data: EmailIdentityProofData,
-    pub reference_block: Option<BlockId>, // the block this identity proof refers to, e.g. an accept, or create block
-}
-
-#[derive(Debug)]
-pub enum CompanyBlockPayload {
-    Create(CompanyCreateBlockData),
-    Update(CompanyUpdateBlockData),
-    SignBill(CompanySignCompanyBillBlockData),
-    InviteSignatory(CompanyInviteSignatoryBlockData),
-    SignatoryAcceptInvite(CompanySignatoryAcceptInviteBlockData),
-    SignatoryRejectInvite(CompanySignatoryRejectInviteBlockData),
-    RemoveSignatory(CompanyRemoveSignatoryBlockData),
-    IdentityProof(CompanyIdentityProofBlockData),
-}
-
-impl Block for CompanyBlock {
-    type OpCode = CompanyOpCode;
-    type BlockDataToHash = CompanyBlockDataToHash;
-
-    fn id(&self) -> BlockId {
-        self.id
-    }
-
-    fn timestamp(&self) -> Timestamp {
-        self.timestamp
-    }
-
-    fn op_code(&self) -> &Self::OpCode {
-        &self.op_code
-    }
-
-    fn plaintext_hash(&self) -> &Sha256Hash {
-        &self.plaintext_hash
-    }
-
-    fn hash(&self) -> &Sha256Hash {
-        &self.hash
-    }
-
-    fn previous_hash(&self) -> &Sha256Hash {
-        &self.previous_hash
-    }
-
-    fn data(&self) -> &[u8] {
-        &self.data
-    }
-
-    fn signature(&self) -> &SchnorrSignature {
-        &self.signature
-    }
-
-    fn public_key(&self) -> &PublicKey {
-        &self.public_key
-    }
-
-    fn validate(&self) -> bool {
-        true
-    }
-
-    /// We validate the plaintext hash against the plaintext data from the CompanyBlockData wrapper
-    fn validate_plaintext_hash(&self, private_key: &secp256k1::SecretKey) -> bool {
-        match from_slice::<CompanyBlockData>(self.data()) {
-            Ok(data_wrapper) => match crypto::decrypt_ecies(&data_wrapper.data, private_key) {
-                Ok(decrypted) => self.plaintext_hash() == &Sha256Hash::from_bytes(&decrypted),
-                Err(e) => {
-                    error!(
-                        "Decrypt Error while validating plaintext hash for id {}: {e}",
-                        self.id()
-                    );
-                    false
-                }
-            },
-            Err(e) => {
-                error!(
-                    "Wrapper Deserialize Error while validating plaintext hash for id {}: {e}",
-                    self.id()
-                );
-                false
-            }
-        }
-    }
-
-    fn get_block_data_to_hash(&self) -> Self::BlockDataToHash {
-        CompanyBlockDataToHash {
-            company_id: self.company_id.clone(),
-            id: self.id(),
-            plaintext_hash: self.plaintext_hash().to_owned(),
-            previous_hash: self.previous_hash().to_owned(),
-            data: self.data().to_owned(),
-            timestamp: self.timestamp(),
-            public_key: self.public_key().to_owned(),
-            signatory_node_id: self.signatory_node_id.clone(),
-            op_code: self.op_code().to_owned(),
-        }
-    }
-}
-
-impl CompanyBlock {
-    /// Create a new block and sign it with an aggregated key, combining the identity key of the
-    /// signer and the company key
-    fn new(
-        company_id: NodeId,
-        id: BlockId,
-        previous_hash: Sha256Hash,
-        data: Vec<u8>,
-        op_code: CompanyOpCode,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-        plaintext_hash: Sha256Hash,
-    ) -> Result<Self> {
-        // The order here is important: identity -> company
-        let keys: Vec<secp256k1::SecretKey> = vec![
-            identity_keys.get_private_key(),
-            company_keys.get_private_key(),
-        ];
-        let signatory_node_id = NodeId::new(identity_keys.pub_key(), company_id.network());
-        let aggregated_public_key = SchnorrSignature::get_aggregated_public_key(&keys)?;
-        let hash = Self::calculate_hash(CompanyBlockDataToHash {
-            company_id: company_id.clone(),
-            id,
-            plaintext_hash: plaintext_hash.clone(),
-            previous_hash: previous_hash.clone(),
-            data: data.clone(),
-            timestamp,
-            public_key: aggregated_public_key.to_owned(),
-            signatory_node_id: signatory_node_id.clone(),
-            op_code: op_code.clone(),
-        })?;
-        let signature = SchnorrSignature::sign_aggregated(&hash, &keys)?;
-
-        Ok(Self {
-            company_id,
-            id,
-            plaintext_hash,
-            hash,
-            timestamp,
-            previous_hash,
-            signature,
-            public_key: aggregated_public_key,
-            signatory_node_id,
-            data,
-            op_code,
-        })
-    }
-
-    pub fn create_block_for_create(
-        company_id: NodeId,
-        genesis_hash: Sha256Hash,
-        company: &CompanyCreateBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let company_bytes = to_vec(company)?;
-        let plaintext_hash = Self::calculate_plaintext_hash(company)?;
-        // encrypt data using company pub key
-        let encrypted_data = crypto::encrypt_ecies(&company_bytes, &company_keys.pub_key())?;
-
-        let key_bytes = to_vec(&company_keys.get_private_key_string())?;
-        // encrypt company keys using creator's identity pub key
-        let encrypted_key = base58::encode(&crypto::encrypt_ecies(
-            &key_bytes,
-            &identity_keys.pub_key(),
-        )?);
-
-        let data = CompanyBlockData {
-            data: encrypted_data,
-            key: Some(encrypted_key),
-        };
-        let serialized_data = to_vec(&data)?;
-
-        Self::new(
-            company_id.to_owned(),
-            BlockId::first(),
-            genesis_hash,
-            serialized_data,
-            CompanyOpCode::Create,
-            identity_keys,
-            company_keys,
-            timestamp,
-            plaintext_hash,
-        )
-    }
-
-    pub fn create_block_for_update(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanyUpdateBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::Update,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_sign_company_bill(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanySignCompanyBillBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::SignCompanyBill,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_invite_signatory(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanyInviteSignatoryBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        signatory_public_key: &PublicKey, // the signatory's public key
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            Some(signatory_public_key),
-            timestamp,
-            CompanyOpCode::InviteSignatory,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_accept_signatory_invite(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanySignatoryAcceptInviteBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::SignatoryAcceptInvite,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_reject_signatory_invite(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanySignatoryRejectInviteBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::SignatoryRejectInvite,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_remove_signatory(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanyRemoveSignatoryBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::RemoveSignatory,
-        )?;
-        Ok(block)
-    }
-
-    pub fn create_block_for_identity_proof(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &CompanyIdentityProofBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let block = Self::encrypt_data_create_block_and_validate(
-            company_id,
-            previous_block,
-            data,
-            identity_keys,
-            company_keys,
-            None,
-            timestamp,
-            CompanyOpCode::IdentityProof,
-        )?;
-        Ok(block)
-    }
-
-    pub fn get_block_data(&self, company_keys: &BcrKeys) -> Result<CompanyBlockPayload> {
-        let data = self.get_decrypted_block_bytes(company_keys)?;
-        let result: CompanyBlockPayload = match self.op_code {
-            CompanyOpCode::Create => CompanyBlockPayload::Create(from_slice(&data)?),
-            CompanyOpCode::Update => CompanyBlockPayload::Update(from_slice(&data)?),
-            CompanyOpCode::InviteSignatory => {
-                CompanyBlockPayload::InviteSignatory(from_slice(&data)?)
-            }
-            CompanyOpCode::SignatoryAcceptInvite => {
-                CompanyBlockPayload::SignatoryAcceptInvite(from_slice(&data)?)
-            }
-            CompanyOpCode::SignatoryRejectInvite => {
-                CompanyBlockPayload::SignatoryRejectInvite(from_slice(&data)?)
-            }
-            CompanyOpCode::RemoveSignatory => {
-                CompanyBlockPayload::RemoveSignatory(from_slice(&data)?)
-            }
-            CompanyOpCode::SignCompanyBill => CompanyBlockPayload::SignBill(from_slice(&data)?),
-            CompanyOpCode::IdentityProof => CompanyBlockPayload::IdentityProof(from_slice(&data)?),
-        };
-        Ok(result)
-    }
-
-    fn get_decrypted_block_bytes(&self, company_keys: &BcrKeys) -> Result<Vec<u8>> {
-        let block_data: CompanyBlockData = from_slice(&self.data)?;
-        let decrypted_bytes =
-            crypto::decrypt_ecies(&block_data.data, &company_keys.get_private_key())?;
-        Ok(decrypted_bytes)
-    }
-
-    fn encrypt_data_create_block_and_validate<T: borsh::BorshSerialize>(
-        company_id: NodeId,
-        previous_block: &Self,
-        data: &T,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        public_key_for_keys: Option<&PublicKey>,
-        timestamp: Timestamp,
-        op_code: CompanyOpCode,
-    ) -> Result<Self> {
-        let bytes = to_vec(&data)?;
-        let plaintext_hash = Self::calculate_plaintext_hash(data)?;
-        // encrypt data using the company pub key
-        let encrypted_data = crypto::encrypt_ecies(&bytes, &company_keys.pub_key())?;
-
-        let mut key = None;
-
-        // in case there are keys to encrypt, encrypt them using the receiver's identity pub key
-        if op_code == CompanyOpCode::InviteSignatory
-            && let Some(signatory_public_key) = public_key_for_keys
-        {
-            let key_bytes = to_vec(&company_keys.get_private_key_string())?;
-            let encrypted_key =
-                base58::encode(&crypto::encrypt_ecies(&key_bytes, signatory_public_key)?);
-            key = Some(encrypted_key);
-        }
-
-        let data = CompanyBlockData {
-            data: encrypted_data,
-            key,
-        };
-        let serialized_data = to_vec(&data)?;
-
-        let new_block = Self::new(
-            company_id,
-            BlockId::next_from_previous_block_id(&previous_block.id),
-            previous_block.hash.clone(),
-            serialized_data,
-            op_code,
-            identity_keys,
-            company_keys,
-            timestamp,
-            plaintext_hash,
-        )?;
-
-        if !new_block.validate_with_previous(previous_block) {
-            return Err(super::Error::BlockInvalid);
-        }
-        Ok(new_block)
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct CompanyBlockchain {
-    blocks: Vec<CompanyBlock>,
-}
-
-impl Blockchain for CompanyBlockchain {
-    type Block = CompanyBlock;
-
-    fn blocks(&self) -> &Vec<Self::Block> {
-        &self.blocks
-    }
-
-    fn blocks_mut(&mut self) -> &mut Vec<Self::Block> {
-        &mut self.blocks
-    }
-}
-
-impl CompanyBlockchain {
-    /// Creates a new company chain
-    pub fn new(
-        company: &CompanyCreateBlockData,
-        identity_keys: &BcrKeys,
-        company_keys: &BcrKeys,
-        timestamp: Timestamp,
-    ) -> Result<Self> {
-        let genesis_hash = Sha256Hash::from_bytes(company.id.to_string().as_bytes());
-
-        let first_block = CompanyBlock::create_block_for_create(
-            company.id.clone(),
-            genesis_hash,
-            company,
-            identity_keys,
-            company_keys,
-            timestamp,
-        )?;
-
-        Ok(Self {
-            blocks: vec![first_block],
-        })
-    }
-
-    /// Creates a company chain from a vec of blocks
-    pub fn new_from_blocks(blocks_to_add: Vec<CompanyBlock>) -> Result<Self> {
-        match blocks_to_add.first() {
-            None => Err(super::Error::BlockchainInvalid),
-            Some(first) => {
-                if !first.verify() || !first.validate_hash() {
-                    return Err(super::Error::BlockchainInvalid);
-                }
-
-                let chain = Self {
-                    blocks: blocks_to_add,
-                };
-
-                if !chain.is_chain_valid() {
-                    return Err(super::Error::BlockchainInvalid);
-                }
-
-                Ok(chain)
-            }
-        }
-    }
-
-    pub fn get_chain_with_plaintext_block_data(
-        &self,
-        company_keys: &BcrKeys,
-    ) -> Result<Vec<CompanyBlockPlaintextWrapper>> {
-        let mut result = Vec::with_capacity(self.blocks().len());
-        for block in self.blocks.iter() {
-            let plaintext_data_bytes = match block.op_code() {
-                CompanyOpCode::Create => block.get_decrypted_block_bytes(company_keys)?,
-                CompanyOpCode::Update => block.get_decrypted_block_bytes(company_keys)?,
-                CompanyOpCode::InviteSignatory => block.get_decrypted_block_bytes(company_keys)?,
-                CompanyOpCode::SignatoryAcceptInvite => {
-                    block.get_decrypted_block_bytes(company_keys)?
-                }
-                CompanyOpCode::SignatoryRejectInvite => {
-                    block.get_decrypted_block_bytes(company_keys)?
-                }
-                CompanyOpCode::RemoveSignatory => block.get_decrypted_block_bytes(company_keys)?,
-                CompanyOpCode::SignCompanyBill => block.get_decrypted_block_bytes(company_keys)?,
-                CompanyOpCode::IdentityProof => block.get_decrypted_block_bytes(company_keys)?,
-            };
-
-            if block.plaintext_hash != Sha256Hash::from_bytes(&plaintext_data_bytes) {
-                return Err(Error::BlockInvalid);
-            }
-
-            result.push(CompanyBlockPlaintextWrapper {
-                block: block.clone(),
-                plaintext_data_bytes,
-            });
-        }
-
-        // Validate the chain from the wrapper
-        CompanyBlockchain::new_from_blocks(
-            result
-                .iter()
-                .map(|wrapper| wrapper.block.to_owned())
-                .collect::<Vec<CompanyBlock>>(),
-        )?;
-
-        Ok(result)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::protocol::{
-        Country,
-        tests::tests::{
-            bill_id_test, node_id_test, node_id_test_other, private_key_test,
-            signed_identity_proof_test, test_ts, valid_address, valid_optional_address,
-        },
-    };
-
-    fn get_baseline_company_data() -> (NodeId, (CompanyCreateBlockData, BcrKeys)) {
-        (
-            node_id_test(),
-            (
-                CompanyCreateBlockData {
-                    id: node_id_test(),
-                    name: Name::new("some_name").unwrap(),
-                    country_of_registration: Some(Country::AT),
-                    city_of_registration: Some(City::new("Vienna").unwrap()),
-                    postal_address: valid_address(),
-                    email: Email::new("company@example.com").unwrap(),
-                    registration_number: Some(Identification::new("some_number").unwrap()),
-                    registration_date: Some(Date::new("2012-01-01").unwrap()),
-                    proof_of_registration_file: None,
-                    logo_file: None,
-                    creation_time: test_ts(),
-                    creator: node_id_test(),
-                },
-                BcrKeys::from_private_key(&private_key_test()),
-            ),
-        )
-    }
-
-    #[test]
-    fn test_plaintext_hash() {
-        let (_id, (company, company_keys)) = get_baseline_company_data();
-
-        let chain = CompanyBlockchain::new(&company, &BcrKeys::new(), &company_keys, test_ts());
-        assert!(chain.is_ok());
-        assert!(chain.as_ref().unwrap().is_chain_valid());
-        assert!(
-            chain.as_ref().unwrap().blocks()[0]
-                .validate_plaintext_hash(&company_keys.get_private_key())
-        );
-    }
-
-    #[test]
-    fn create_and_check_validity() {
-        let (_id, (company, company_keys)) = get_baseline_company_data();
-
-        let chain = CompanyBlockchain::new(&company, &BcrKeys::new(), &company_keys, test_ts());
-        assert!(chain.is_ok());
-        assert!(chain.as_ref().unwrap().is_chain_valid());
-    }
-
-    #[test]
-    fn multi_block() {
-        let (id, (company, company_keys)) = get_baseline_company_data();
-        let identity_keys = BcrKeys::new();
-
-        let chain = CompanyBlockchain::new(&company, &identity_keys, &company_keys, test_ts());
-        assert!(chain.is_ok());
-        assert!(chain.as_ref().unwrap().is_chain_valid());
-
-        let mut chain = chain.unwrap();
-        let update_block = CompanyBlock::create_block_for_update(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanyUpdateBlockData {
-                name: Some(Name::new("new_name").unwrap()),
-                email: None,
-                postal_address: valid_optional_address(),
-                country_of_registration: None,
-                city_of_registration: None,
-                registration_number: None,
-                registration_date: None,
-                logo_file: None,
-                proof_of_registration_file: None,
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 1,
-        );
-        assert!(update_block.is_ok());
-        chain.try_add_block(update_block.unwrap());
-
-        let bill_block = CompanyBlock::create_block_for_sign_company_bill(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanySignCompanyBillBlockData {
-                bill_id: bill_id_test(),
-                block_id: BlockId::first(),
-                block_hash: Sha256Hash::new("some hash"),
-                operation: BillOpCode::Issue,
-                bill_key: Some(private_key_test()),
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 2,
-        );
-        assert!(bill_block.is_ok());
-        chain.try_add_block(bill_block.unwrap());
-
-        let invite_signatory_block = CompanyBlock::create_block_for_invite_signatory(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanyInviteSignatoryBlockData {
-                invitee: node_id_test(),
-                inviter: node_id_test_other(),
-                t: SignatoryType::Solo,
-            },
-            &identity_keys,
-            &company_keys,
-            &node_id_test().pub_key(),
-            test_ts() + 3,
-        );
-        assert!(invite_signatory_block.is_ok());
-        chain.try_add_block(invite_signatory_block.unwrap());
-
-        let accept_invitation_block = CompanyBlock::create_block_for_accept_signatory_invite(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanySignatoryAcceptInviteBlockData {
-                accepter: node_id_test(),
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 4,
-        );
-        assert!(accept_invitation_block.is_ok());
-        chain.try_add_block(accept_invitation_block.unwrap());
-
-        let reject_invitation_block = CompanyBlock::create_block_for_reject_signatory_invite(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanySignatoryRejectInviteBlockData {
-                rejecter: node_id_test(),
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 5,
-        );
-        assert!(reject_invitation_block.is_ok());
-        chain.try_add_block(reject_invitation_block.unwrap());
-
-        let remove_signatory_block = CompanyBlock::create_block_for_remove_signatory(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanyRemoveSignatoryBlockData {
-                removee: node_id_test(),
-                remover: node_id_test_other(),
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 6,
-        );
-        assert!(remove_signatory_block.is_ok());
-        chain.try_add_block(remove_signatory_block.unwrap());
-
-        let test_signed_identity = signed_identity_proof_test();
-        let identity_proof_block = CompanyBlock::create_block_for_identity_proof(
-            id.clone(),
-            chain.get_latest_block(),
-            &CompanyIdentityProofBlockData {
-                proof: test_signed_identity.0,
-                data: test_signed_identity.1,
-                reference_block: None,
-            },
-            &identity_keys,
-            &company_keys,
-            test_ts() + 7,
-        );
-        assert!(identity_proof_block.is_ok());
-        chain.try_add_block(identity_proof_block.unwrap());
-
-        assert_eq!(chain.blocks().len(), 8);
-        assert!(chain.is_chain_valid());
-
-        let new_chain_from_empty_blocks = CompanyBlockchain::new_from_blocks(vec![]);
-        assert!(new_chain_from_empty_blocks.is_err());
-
-        let blocks = chain.blocks();
-
-        for block in blocks {
-            assert!(block.validate_plaintext_hash(&company_keys.get_private_key()));
-        }
-
-        let new_chain_from_blocks = CompanyBlockchain::new_from_blocks(blocks.to_owned());
-        assert!(new_chain_from_blocks.is_ok());
-        assert!(new_chain_from_blocks.as_ref().unwrap().is_chain_valid());
-
-        let mut_blocks = chain.blocks_mut();
-        mut_blocks[2].hash = Sha256Hash::new("invalidhash");
-        let new_chain_from_invalid_blocks =
-            CompanyBlockchain::new_from_blocks(mut_blocks.to_owned());
-        assert!(new_chain_from_invalid_blocks.is_err());
-    }
+    pub signer_node_id: NodeId,
+    pub op: CompanyOpCode,
+    pub company_keys: BcrKeys,
+    pub invitee: Option<NodeId>,
+    pub removee: Option<NodeId>,
+    pub identity_proof_data: Option<(SignedIdentityProof, EmailIdentityProofData, Option<BlockId>)>,
 }
